blog_server/requirements.txt
coreapi==2.3.3
cryptography==36.0.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
